(!!!!!READ NOTE!!!!!)Added a test project with the modified driver for DDR memory.
NOTE: Running this project on the De1 SoC causes errors which cannot be resolved even when the FPGA is programmed with a different file.

This project allows playback for 39s, but state does not transition to PASS afterwards (from REC as well as PLAY) and the reset button needs to be pressed.

Extensive debugging needs to be conducted before uploading this project to the board.

It is best to only look at the modified driver and find the errors.